Updates to z3-object-with-handle -> string fns
To simplify code related to printing z3-object-with-handle subtypes,
created a new generic method z3-object-to-string and implemented
describe-object and print-object for z3-object-with-handle based on that
function.

z3-object-with-handle subclasses should now define a version of the
z3-object-to-string method for that specific subclass. After doing so
that subclass will get the nicer describe and printing functionality for
free.