Validate parameters before setting them
set-params now validates the given params object against the solver's
param-descrs to avoid situations where an invalid parameter setting
could cause any further interaction with the solver to emit an error.

Note that the validation only detects "obvious" problems, e.g. a
parameter being set to a value of the wrong type. Some other problems -
like setting the string_solver parameter to a symbol that does not
designate a valid string solver - will not be detected by the validation
and instead will only be detected later when the solver is used.