Change the way that algebraic numbers are output
Algebraic numbers (e.g. irrational numbers) are now output as Lisp
floating point numbers by default, though this can be changed via a
parameter. This necessitated the addition ot the parse-float library, as
Common Lisp does not provide a built-in mechanism for parsing a string
as a float (outside of the reader, the use of which for this purpose is
problematic).