Introduce persistent env for variables, fix function calls
Introduce a persistent environment for variables that can be used in
assertions. Variables are added to this environment either by previous
z3-assert forms, or by using declare-const or declare-fun. The env is
appropriately synced with changes to the assertion stack (e.g.
push/pop).

Disallow multiple variables with the same name but different sorts being
defined, either in the same z3-assert or across them in the same
assertion stack. This removes ambiguity from the alist that
get-model-as-assignment returns.

Fix the handling of uninterpreted function calls. Previously these were
made using _ rather than normal function call syntax, but this is not
consistent with SMT-LIB2. Now, _ is not supported and uninterpreted
function calls look the same as interpreted function calls.

Updated several examples to reflect the changes made.