Update autogen scripts, regenerate CFFI files
Rename some of the CFFI foreign types to make it clearer what is "low
level".

Update the scripts that are used to regenerate the files needed to
perform CFFI bindings with Z3, and regenerate the CFFI files based on Z3
4.8.15 (as this seems to be the oldest version such that I don't see any
issues when running the interface with Z3 4.13.4, the current release).
Note that I manually removed the Z3_literals type as it was removed
after 4.8.15, and was deprecated prior to that.

Update the README to describe the versions of Z3 that we expect this to
work with, and to say that versions outside of this range may work if
the user regenerates the CFFI files (and to point to the script that
does so, which has comments describing how to use it and why one might
want to use it).